requirements.txt
streamlit
python-docx
langchain
pdfplumber
openai
deepseek-transformers
qwen
grok-api
requests